requirements.txt
astroid==3.3.10
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
dill==0.4.0
fonttools==4.58.0
isort==6.0.1
kiwisolver==1.4.8
matplotlib==3.10.3
mccabe==0.7.0
mysql-connector-python==9.3.0
numpy==2.2.5
packaging==25.0
pandas==2.2.3
pillow==11.2.1
platformdirs==4.3.8
pylint==3.3.7
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
seaborn==0.13.2
six==1.17.0
tomli==2.2.1
tomlkit==0.13.2
typing_extensions==4.13.2
tzdata==2025.2
